refactor EISBuilder, EISProblem, adds python costs
