+++ pybop/__init__.py
@@ -106,12 +106,12 @@ from .costs.pybamm_cost import (
 )
 from .costs.pybamm_sum_squared_error import PybammSumSquaredError
 
-from .costs.base_cost import BaseCost
+from .costs.base_cost_old import BaseCost
 from .costs.fitting_costs import (
     FittingCost,
     ObserverCost,
 )
-from .costs.error_measures import (
+from .costs.error_measures_old import (
     RootMeanSquaredError,
     MeanAbsoluteError,
     MeanSquaredError,
@@ -119,6 +119,7 @@ from .costs.error_measures import (
     Minkowski,
     SumOfPower,
 )
+from .costs.error_measures import NewMeanSquaredError
 from .costs.design_costs import (
     DesignCost,
     GravimetricEnergyDensity,
