+++ pybop/builders/pybamm_eis.py
@@ -1,20 +1,37 @@
-from pybop import Parameter, PybammEISProblem, builders
+from typing import Callable
+
+import pybamm
+
+from pybop import PybammEISProblem, builders
 from pybop._pybamm_eis_pipeline import PybammEISPipeline
-from pybop.costs.pybamm_cost import PybammCost
 
 
-class PybammEIS(builders.Pybamm):
+class PybammEIS(builders.BaseBuilder):
     def __init__(self):
         super().__init__()
         self.domain = "Frequency [Hz]"
 
-    def add_cost(self, cost: PybammCost, weight: float = 1.0) -> None:
-        self._costs.append(cost)
-        self._cost_names.append(cost.variable_name())
-        self._cost_weights.append(weight)
+    def set_simulation(
+        self,
+        model: pybamm.BaseModel,
+        parameter_values: pybamm.ParameterValues = None,
+        initial_state: tuple = None,
+    ) -> None:
+        """
+        Adds a simulation for the optimisation problem.
+        """
+        self._model = model.new_copy()
+        self._initial_state = initial_state
+        self._parameter_values = parameter_values or model.default_parameter_values
+        self._solver = pybamm.CasadiSolver()
 
-    def add_parameter(self, parameter: Parameter) -> None:
-        self._pybop_parameters.add(parameter)
+    def add_cost(self, cost: Callable, weight: float = 1.0) -> None:
+        """
+        Add domain weighting option here? Or maybe, set_simulation? Or maybe dataset
+        Options: "equal"(default), "domain", or a custom numpy array.
+        """
+        self._costs.append(cost)
+        self._cost_weights.append(weight)
 
     def build(self) -> PybammEISProblem:
         """
@@ -38,7 +55,7 @@ class PybammEIS(builders.Pybamm):
                 "Number of cost weights and the number of costs do not match"
             )
 
-        if self._pybamm_model is None:
+        if self._model is None:
             raise ValueError("A Pybamm model needs to be provided before building.")
 
         if self._costs is None:
@@ -47,18 +64,14 @@ class PybammEIS(builders.Pybamm):
         if self._dataset is None:
             raise ValueError("A dataset must be provided before building.")
 
-        # Proceed to building the pipeline
-        model = self._pybamm_model
+        # Proceed to build the pipeline
+        model = self._model
         param = self._parameter_values
         pybop_parameters = self._pybop_parameters
 
         # Build pybamm if not already built
         if not model._built:  # noqa: SLF001
             model.build_model()
-
-        # add costs
-        for cost in self._costs:
-            cost.add_to_model(model, param)
 
         # Construct the pipeline
         pipeline = PybammEISPipeline(
@@ -75,6 +88,7 @@ class PybammEIS(builders.Pybamm):
         return PybammEISProblem(
             pybamm_pipeline=pipeline,
             pybop_params=self._pybop_parameters,
-            cost_names=self._cost_names,
+            costs=self._costs,
             cost_weights=self._cost_weights,
+            fitting_data=self._dataset["Impedance"],
         )
